refact : interview/screen-result 구조 변경
- content-detail : result 결과
- feedback-modal : 피드백 모달
- footer-panel : 푸터 액션 모음
- record-navigation : 결과 넘버 전환
- record-player : 결과 출력 플레이어
