+++ client/src/features/interview/screen-result/model/index.ts
@@ -1,2 +1,3 @@
 export { default as useDownloadHandler } from './useDownloadHandler';
 export { default as useIntvFeedback } from './useIntvFeedback';
+export * from './types';
