+++ client/src/features/interview/screen-result/model/types.ts
@@ -1,0 +1,12 @@
+type ResultStateProps = {
+    video: Blob[];
+    recordContents: {
+      script: string;
+      timer: string;
+      voiceScript: string;
+    }[];
+    question: string[];
+    counter: number;
+  };
+
+export type { ResultStateProps };
