+++ client/src/features/interview/screen-result/ui/feedback-modal/index.ts
@@ -1,0 +1,1 @@
+export { default as IntvFeedbackModal } from './IntvFeedbackModal';
