+++ client/src/features/interview/screen-result/ui/footer-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as FooterActionPanel } from './FooterActionPanel';
