+++ client/src/features/interview/screen-result/ui/record-navigation/RecordNavigation.test.tsx
