+++ client/src/features/interview/screen-result/ui/record-navigation/index.ts
@@ -1,0 +1,1 @@
+export { default as RecordNavigation } from './RecordNavigation';
