+++ client/src/features/interview/screen-result/ui/record-player/RecordPlayer.test.tsx
