+++ client/src/features/interview/screen-result/ui/record-player/index.ts
@@ -1,0 +1,1 @@
+export { default as RecordPlayer } from './RecordPlayer';
